pyproject.toml
[project]
name = "py-secscan"
version = "0.1.0"
description = "PySecScan is the tool that allows you to perform security scans in your Python project through a simple YAML configuration."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cyclonedx-bom>=5.1.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.8.4",
]

[project.scripts]
py_secscan = "py_secscan:main"

[tool.uv.workspace]
members = ["py-secscan"]

[tool.uv]
package = true
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]
